requirements.txt
fastapi==0.118.1
pydantic==2.12.0
pydantic_core==2.41.1
uvicorn==0.37.0
PyJWT==2.10.1